Move the EventIntegrator to a dedicated file.
